requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
h11==0.13.0
idna==3.3
outcome==1.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
robotframework==5.0.1
robotframework-pythonlibcore==3.0.0
robotframework-seleniumlibrary==7.0.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0